requirements.txt
flask>=2.0.0
httpx>=0.21.0
python-dotenv>=0.19.0
